Part2:Front-end Index Page Setting v2
